requirements.txt
git+https://github.com/Rapptz/discord.py
pillow
pyjokes
requests
